upbeat/requirements.txt
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.0
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
numpy==2.1.1
packaging==24.1
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
# scikit-learn==1.5.2
# scikit-learn==1.2.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
uritemplate==4.1.1
